requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asn1crypto==1.5.1
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
fastapi==0.125.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pg8000==1.31.5
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scramp==1.4.6
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
